pyproject.toml
[tool.poetry]
name = "offstreet-parking"
version = "0.1.0"
description = "A tool to fetch and upload offstreet parking data to NIPKaart."
authors = ["NIPKaart <hello@nipkaart.nl>"]
maintainers = ["Klaas Schoute <hello@student-techlife.com>"]
license = "MIT"
readme = "README.md"
homepage = "https://github.com/nipkaart/offstreet-parking"
repository = "https://github.com/nipkaart/offstreet-parking"
documentation = "https://github.com/nipkaart/offstreet-parking"
keywords = ["offstreet", "parking", "nipkaart", "locations", "api"]
packages = [
  { include = "app" },
]

[tool.poetry.dependencies]
pymysql = "^1.1.0"
python = "^3.10"
python-dotenv = "^1.0.0"
pytz = "^2023.3"

[tool.poetry.group.cities.dependencies]
hamburg = "^1.0.1"
odp-amsterdam = "^5.1.1"

[tool.poetry.group.dev.dependencies]
black = "23.7.0"
blacken-docs = "1.15.0"
pre-commit = "3.3.3"
pre-commit-hooks = "4.4.0"
pylint = "2.17.4"
ruff = "0.0.280"
yamllint = "1.32.0"

[tool.pylint.BASIC]
good-names = [
  "_",
  "ex",
  "fp",
  "i",
  "id",
  "j",
  "k",
  "on",
  "Run",
  "T",
]

[tool.pylint."MESSAGES CONTROL"]
disable= [
  "too-few-public-methods",
  "too-many-arguments",
  "duplicate-code",
  "format",
  "unsubscriptable-object",
  "broad-except",
]

[tool.pylint.SIMILARITIES]
ignore-imports = true

[tool.pylint.FORMAT]
max-line-length = 88

[tool.pylint.DESIGN]
max-attributes = 20

[tool.ruff]
select = ["ALL"]
ignore = [
  "ANN101", # Self... explanatory
  "ANN401", # Opinioated warning on disallowing dynamically typed expressions
  "D203", # Conflicts with other rules
  "D213", # Conflicts with other rules
  "D417", # False positives in some occasions
  "PLR2004", # Just annoying, not really useful
  "T201", # Allow the use of print()
]

[tool.ruff.flake8-pytest-style]
mark-parentheses = false
fixture-parentheses = false

[build-system]
build-backend = "poetry.core.masonry.api"
requires = ["poetry-core>=1.0.0"]
